+++ Classes/Command.js
@@ -1,7 +1,7 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 
 module.exports = class Command {
-	constructor(nameInput, descriptionInput, optionsInput) {
+	constructor(nameInput, descriptionInput, optionsInput, subcommandsInput) {
 		this.name = nameInput;
 		this.description = descriptionInput;
 		this.data = new SlashCommandBuilder()
@@ -17,6 +17,25 @@ module.exports = class Command {
 				if (choiceEntries.length) {
 					built.addChoices(Object.entries(option.choices));
 				}
+				return built;
+			})
+		})
+		subcommandsInput.forEach(subcommand => {
+			this.data.addSubcommand(built => {
+				built.setName(subcommand.name).setDescription(subcommand.description);
+				subcommand.optionsInput.forEach(option => {
+					built[`add${option.type}Option`](subBuilt => {
+						subBuilt.setName(option.name).setDescription(option.description).setRequired(option.required);
+						if (option.choices === null || option.choices === undefined) {
+							throw `Error: ${this.nameInput} (${descriptionInput}) ${option.type} Option was nullish.`;
+						}
+						let choiceEntries = Object.entries(option.choices);
+						if (choiceEntries.length) {
+							subBuilt.addChoices(Object.entries(option.choices));
+						}
+						return subBuilt;
+					})
+				})
 				return built;
 			})
 		})