+++ Source/Commands/club-config.js
@@ -2,7 +2,7 @@ const Command = require('../../Classes/Command.js');
 
 const COLORS = ["WHITE", "AQUA", "GREEN", "BLUE", "YELLOW", "PURPLE", "LUMINOUS_VIVID_PINK", "FUCHSIA", "GOLD", "ORANGE", "RED", "GREY", "NAVY", "DARK_AQUA", "DARK_GREEN", "DARK_BLUE", "DARK_PURPLE", "DARK_VIVID_PINK", "DARK_GOLD", "DARK_ORANGE", "DARK_RED", "DARK_GREY", "BLURPLE", "GREYPLE", "RANDOM"];
 
-let options = [
+const options = [
 	{ type: "String", name: "name", description: "What to call the club", required: false, choices: {} },
 	{ type: "String", name: "description", description: "Text shown in the channel topic", required: false, choices: {} },
 	{ type: "String", name: "game", description: "The text to set as the club game", required: false, choices: {} },
@@ -14,7 +14,8 @@ let options = [
 		}, {})
 	}
 ];
-module.exports = new Command("club-config", "(club leader or moderator) Configure a club's information", options);
+const subcomands = [];
+module.exports = new Command("club-config", "(club leader or moderator) Configure a club's information", options, subcomands);
 
 let isModerator, getClubs, updateClub, updateClubDetails;
 module.exports.initialize = function (helpers) {