+++ Source/Commands/club-promote-leader.js
@@ -1,7 +1,8 @@
 const Command = require('../../Classes/Command.js');
 
-let options = [{ type: "User", name: "user", description: "The user's mention", required: true, choices: {} }];
-module.exports = new Command("club-promote-leader", "(club leader or moderator) Promote another user to club leader", options);
+const options = [{ type: "User", name: "user", description: "The user's mention", required: true, choices: {} }];
+const subcommands = [];
+module.exports = new Command("club-promote-leader", "(club leader or moderator) Promote another user to club leader", options, subcommands);
 
 let isModerator, getClubs, updateClub, updateClubDetails;
 module.exports.initialize = function (helpers) {