+++ Source/Commands/embed-add-field.js
@@ -1,12 +1,13 @@
 const Command = require('../../Classes/Command.js');
 
-let options = [
+const options = [
 	{ type: "String", name: "message-id", description: "The id of the embed's message", required: true, choices: {} },
 	{ type: "String", name: "header", description: "The header text", required: true, choices: {} },
 	{ type: "String", name: "text", description: "The body text", required: true, choices: {} },
 	{ type: "Boolean", name: "inline", description: "Whether to show the field in-line with previous embed fields", required: false, choices: {} }
 ];
-module.exports = new Command("embed-add-field", "(moderator) Add a custom embed field", options);
+const subcomands = [];
+module.exports = new Command("embed-add-field", "(moderator) Add a custom embed field", options, subcomands);
 
 let customEmbeds, isModerator;
 module.exports.initialize = function (helpers) {