+++ Source/Commands/embed-set-message.js
@@ -1,10 +1,11 @@
 const Command = require('../../Classes/Command.js');
 
-let options = [
+const options = [
 	{ type: "String", name: "message-id", description: "The id of the embed's message", required: true, choices: {} },
 	{ type: "String", name: "text", description: "The text to put in the message (above the embed)", required: true, choices: {} }
 ];
-module.exports = new Command("embed-set-message", "(moderator) Set a custom embed's message content", options);
+const subcomands = [];
+module.exports = new Command("embed-set-message", "(moderator) Set a custom embed's message content", options, subcomands);
 
 let customEmbeds, isModerator;
 module.exports.initialize = function (helpers) {