+++ Source/Commands/embed-set-thumbnail.js
@@ -1,10 +1,11 @@
 const Command = require('../../Classes/Command.js');
 
-let options = [
+const options = [
 	{ type: "String", name: "message-id", description: "The id of the embed's message", required: true, choices: {} },
 	{ type: "String", name: "url", description: "The picture url for the thumbnail", required: true, choices: {} },
 ];
-module.exports = new Command("embed-set-thumbnail", "(moderator) Set a custom embed's thumbnail", options);
+const subcomands = [];
+module.exports = new Command("embed-set-thumbnail", "(moderator) Set a custom embed's thumbnail", options, subcomands);
 
 let customEmbeds, isModerator;
 module.exports.initialize = function (helpers) {