+++ Source/Commands/join.js
@@ -1,10 +1,11 @@
 const Command = require('../../Classes/Command.js');
 
-// can't use channel mention because users can't mention channels that are invisible to them (even by constructing the mention manually)
-let options = [
+const options = [
+	// can't use channel mention because users can't mention channels that are invisible to them (even by constructing the mention manually)
 	{ type: "String", name: "channel", description: "The name/id of the topic or club to join", required: true, choices: {} },
 ];
-module.exports = new Command("join", "Join a topic or club", options);
+const subcomands = [];
+module.exports = new Command("join", "Join a topic or club", options, subcomands);
 
 let joinChannel, findTopicID;
 module.exports.initialize = function (helpers) {