+++ Source/Commands/timestamp.js
@@ -1,12 +1,13 @@
 const Command = require('../../Classes/Command.js');
 
-let options = [
+const options = [
 	{ type: "String", name: "start", description: "The timestamp to start from (default: now)", required: false, choices: {} },
 	{ type: "Number", name: "days-from-start", description: "86400 seconds", required: false, choices: {} },
 	{ type: "Number", name: "hours-from-start", description: "3600 seconds", required: false, choices: {} },
 	{ type: "Number", name: "minutes-from-start", description: "60 seconds", required: false, choices: {} }
 ];
-module.exports = new Command("timestamp", "Calculate the unix timestamp for a moment in time, which Discord displays with timezones applied", options);
+const subcomands = [];
+module.exports = new Command("timestamp", "Calculate the unix timestamp for a moment in time, which Discord displays with timezones applied", options, subcomands);
 
 let timeConversion;
 module.exports.initialize = function (helpers) {